requirements.txt
streamlit
ollama
pillow
python-dotenv
